+++ Truck.cs
@@ -37,7 +37,7 @@ namespace PowerLine_TestTask
         }
         public decimal AverageFuelConsumption { get; }
 
-        public Truck(string type, int maxSpeed, int fuel, int aveFuelCons)
+        public Truck(string type, int maxSpeed, int fuel, decimal aveFuelCons)
         {
             if (maxSpeed <= 0 || fuel <= 0 || aveFuelCons <= 0)
                 throw new ArgumentException("Params can't be negative or null");
